requirements.txt
langgraph
dspy-ai
google-genai
chromadb
gitpython
pydantic
python-dotenv
